pyproject.toml
[project]
name = "agent-engine"
version = "0.1.0"
description = "Pydantic AI Agent Platform on VertexAI Agent Engine"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Platform Team"}
]
keywords = ["ai", "agent", "pydantic", "vertexai", "gemini"]

dependencies = [
    "pydantic-ai-slim[google]>=1.51.0",
    "google-cloud-aiplatform[agent_engines]>=1.78.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # Phase 2: Observability
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "opentelemetry-exporter-gcp-monitoring>=1.6.0",
    "google-cloud-trace>=1.11.0",
    "google-cloud-logging>=3.8.0",
    "google-cloud-monitoring>=2.18.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    # Phase 3: CI/CD & Security
    "bandit>=1.7.0",
    "radon>=6.0.0",
    # Phase 3: Version Management
    "pyyaml>=6.0.0",
    "semver>=3.0.0",
    # Phase 3: Evaluation & Performance
    "pytest-benchmark>=4.0.0",
    "locust>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_engine"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]  # Allow imports after sys.path modification

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/agent_engine"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.coverage.html]
directory = "coverage_html"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert warnings in non-test code
severity = "ll"  # Only report medium and high severity

[tool.bandit.assert_used]
skips = ["*_test.py", "test_*.py"]
